requirements.txt
gradio
pywebview

pillow
opencv-python

torch
torchvision
einops
timm
tqdm
onnx
onnxruntime
transformers

kornia
yolov8-onnx

numpy
pandas
pySmartDL